requirements.txt
torch
torchvision
wandb
tqdm
scikit-learn
PyYAML
numpy
timm